Fix for boxing/unboing in initializers
