+++ sources/net.sf.j2s.core/src/j2s/swingjs/Java2ScriptSwingJSCompiler.java
@@ -27,7 +27,7 @@ import j2s.core.Java2ScriptCompiler;
  */
 public class Java2ScriptSwingJSCompiler extends Java2ScriptCompiler {
 
-	private int nResources, nSources, nJS, nHTML;
+	private int nSources, nJS, nHTML;
 
 	private Map<String, String> htMethodsCalled;
 	private List<String> lstMethodsDeclared;
@@ -283,7 +283,7 @@ public class Java2ScriptSwingJSCompiler extends Java2ScriptCompiler {
 			addHTML(visitor.getAppList(false), siteFolder, htmlTemplate, false);
 			String packageName = visitor.getMyPackageName();
 			if (packageName != null) {
-				nResources += checkCopiedResources(packageName, sourceLocation, j2sPath);
+				copyAllResources(packageName, sourceLocation);
 			}
 			return true;
 		} catch (Throwable e) {
