+++ sources/net.sf.j2s.java.core/site-resources_4.2/jsmol/j2s/java/util/Arrays.js
@@ -202,6 +202,4 @@ function(a){
 return this.indexOf(a)!=-1;
 },"~O");
 /*eoif*/})();
-//Clazz.defineStatics(c$,
-//"INSERTIONSORT_THRESHOLD",7);
 });
