+++ sources/net.sf.j2s.java.core/site-resources_4.2/jsmol/j2s/java/util/HashMap.js
@@ -510,6 +510,4 @@ return a;
 /*eoif*/})();
 };
 /*eoif*/})();
-//Clazz.defineStatics(c$,
-//"DEFAULT_SIZE",16);
 });
