+++ sources/net.sf.j2s.java.core/src_4.2/test/ITest.java
@@ -1,6 +1,9 @@
 package test;
 
-public interface ITest {
+public interface ITest { 
+//   static String stest = new String("stest2");
+//   final static String atest = new String("atest");
+//   final static String btest = "btest";
    public String itest();	
 	
 }
