+++ sources/net.sf.j2s.java.core/src_4.2/test/J2SIgnoreImport.java
@@ -1,0 +1,7 @@
+package test;
+
+public @interface J2SIgnoreImport {
+
+	Class<?>[] value();
+
+}
