+++ sources/net.sf.j2s.java.core/src_4.2/test/J2SRequireImport.java
@@ -1,0 +1,7 @@
+package test;
+
+public @interface J2SRequireImport {
+
+	Class<?>[] value();
+
+}
