+++ sources/net.sf.j2s.java.core/src_4.2/test/Test.java
@@ -1,10 +1,41 @@
 package test;
 
-public class Test {
+public class Test implements ITest {
 	
-	static int i = 1;
+	static int i = '2';
+	final static String s = "test " + "here";
+	final static String s1 = new String("test5");
+	static String s3 = new String("test8");
+	static String s2 = "s2";
+	static char c = 'c';
+	
+	
+	int j = 'j';
+	String ss = "sst";
+	String st = new String("st");
+	char d = 'd';	
+	char e = (char) ('1' + d);
+	char f = (char) ('1' + '\1');
+	
+	char cChar2char = new Character('w'); // fixed
+	
+	Character char2Char = 'g';
+
+	int iChar2int = new Character('h'); // fixed
+	
+	static char Char2char_static = new Character('\121');
+		Character int2Char = 121; // fixed
+	
+	Character short2Char = (short) 120; //fixed
+	
+	static Character int2Char_static = 121; // fixed
+	static int iChar2int_static = new Character('c');
+
+	Integer int2Int = 122;
+	int iInt2int = Integer.valueOf("200"); // fixed
 	
 	public Test(){
+		int iInt2int = Integer.valueOf("201"); // fixed
 		
 	}
 
@@ -48,6 +79,17 @@ public class Test {
 		
 		
 		
+	}
+
+	@Override
+	public String itest() {
+//		String a = atest;
+//		String b = btest;
+		String c = "c";
+		String d = new String("d");
+		String e = s1;
+		String f = s;
+		return "OK1";//this.stest;
 	}
